requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
requests==2.28.1
FPDF==1.7.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
itsdangerous==2.1.2
Flask-WTF==1.1.1
gunicorn==20.1.0